# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ClearDeadQueueRequest:
    properties:
      queue_name:
        example: default
        type: string
      worker_name:
        example: my-worker
        type: string
    required:
    - worker_name
    type: object
  dto.ClearDeadQueueResponse:
    properties:
      cleared_queues:
        items:
          type: string
        type: array
      status:
        example: ok
        type: string
      total_deleted:
        example: 50
        type: integer
      worker_name:
        example: my-worker
        type: string
    type: object
  dto.ClearQueueRequest:
    properties:
      queue_name:
        example: default
        type: string
      worker_name:
        example: my-worker
        type: string
    required:
    - worker_name
    type: object
  dto.ClearQueueResponse:
    properties:
      cleared_queues:
        items:
          type: string
        type: array
      status:
        example: ok
        type: string
      total_deleted:
        example: 100
        type: integer
      worker_name:
        example: my-worker
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: 错误信息
        type: string
    type: object
...
  dto.QueueStatsResponse:
    properties:
      queues:
        items: {}
        type: array
      worker_name:
        example: my-worker
        type: string
    type: object
...
host: localhost:28080
info:
  contact:
    name: TaskPM Support
  description: 分布式任务管理系统 - 基于 Asynq 和 PostgreSQL 的任务调度平台
  license:
    name: MIT
  title: TaskPM API
  version: 1.0.0
paths:
...
  /queues/clear:
    post:
      consumes:
      - application/json
      description: 清空指定 Worker 的队列（删除所有待处理任务）
      parameters:
      - description: 清空队列请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ClearQueueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ClearQueueResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 清空队列
      tags:
      - Queues
  /queues/clear-dead:
    post:
      consumes:
      - application/json
      description: 清空指定 Worker 的死信队列（已达最大重试次数的失败任务）
      parameters:
      - description: 清空死信队列请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ClearDeadQueueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ClearDeadQueueResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 清空死信队列
      tags:
      - Queues
  /queues/stats:
    get:
      description: 获取指定 Worker 的所有队列状态统计
      parameters:
      - description: Worker 名称
        in: query
        name: worker_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.QueueStatsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 查询队列状态
      tags:
      - Queues
...
schemes:
- http
- https
swagger: "2.0"